muOS Internals - Added check for manually launched content - Added startup removal of update scripts - Added wait auto loader for last and resume game state for RetroArch - Fixed network SSID in global configuration - Removed soundfont archive (default set in base)

+++ script/mux/launch.sh
@@ -129,7 +129,9 @@ cat /dev/zero >"$DC_SCR_DEVICE" 2>/dev/null
 cat /dev/zero >"$DC_SCR_DEVICE" 2>/dev/null
 
 if [ "$GC_GEN_STARTUP" = last ] || [ "$GC_GEN_STARTUP" = resume ]; then
-	echo launcher >"$ACT_GO"
+	if [ ! -e "/tmp/manual_launch" ]; then
+		echo launcher >"$ACT_GO"
+	fi
 fi
 
 killall -q "$GPTOKEYB_BIN" "$EVSIEVE_BIN"
